sys-devel/llvm: Add Bash version of llvm-config for cross-compiling
This is a poor man's rewrite of llvm-config in Bash, which is needed
because llvm-config is normally a native binary and cannot be used
when cross-compiling.

It supports most of the llvm-config flags and will return largely
accurate answers. The main difference is that it doesn't support
filtering by component so it will always return all of the LLVM
libraries. This would cause overlinking but Gentoo avoids this by
default with -Wl,--as-needed.

To use this, llvm.eclass should add
${ESYSROOT}/usr/lib/llvm/${SLOT}/bin/cross to the PATH instead of the
parent directory but it should only do so when cross-compiling.

Tested with LLVM 7 and 8 against Mesa.

Package-Manager: Portage-2.3.66, Repoman-2.3.12
Signed-off-by: James Le Cuirot <chewi@gentoo.org>